ML updates to training models
