+++ nhlPredict/src/ML/model_training/Predict_SOG_v2-Manual-with-gameDate.py
@@ -9,6 +9,7 @@
 # COMMAND ----------
 
 import mlflow
+import databricks.automl_runtime
 
 target_col = "player_Total_shotsOnGoal"
 time_col = "gameDate"
@@ -20,13 +21,16 @@ time_col = "gameDate"
 
 # COMMAND ----------
 
+import mlflow
+import os
+import uuid
 import shutil
 import pandas as pd
 
 df_loaded = spark.table("lr_nhl_demo.dev.SOG_features_v2")
 
 # Preview data
-display(df_loaded.limit(5))
+display(df_loaded.head(5))
 
 # COMMAND ----------
 
@@ -60,16 +64,9 @@ display(df_loaded.limit(5))
 
 # COMMAND ----------
 
-cols_to_remove = [
-    "DAY",
-    "HOME",
-    "AWAY",
-    "gameId",
-    "playerId",
-    "shooterName",
-    "home_or_away",
-    time_col,
-]
+from pyspark.sql.types import StringType
+
+cols_to_remove = ["DAY", "HOME", "AWAY", "gameId", "playerId", "shooterName", "home_or_away"]
 
 # Identify numerical and categorical columns
 numerical_cols = [
@@ -91,8 +88,8 @@ print(numerical_cols)
 from databricks.automl_runtime.sklearn.column_selector import ColumnSelector
 
 supported_cols = categorical_cols + [col for col in numerical_cols if col != target_col]
-# supported_cols = list(set(supported_cols + [time_col]) - set(cols_to_remove))
-supported_cols = list(set(supported_cols) - set(cols_to_remove))
+supported_cols = list(set(supported_cols + [time_col]) - set(cols_to_remove))
+# supported_cols = list(set(supported_cols) - set(cols_to_remove))
 
 col_selector = ColumnSelector(supported_cols)
 
@@ -132,6 +129,7 @@ target_col in numerical_cols
 
 # COMMAND ----------
 
+from pandas import Timestamp
 from sklearn.pipeline import Pipeline
 
 from databricks.automl_runtime.sklearn import DatetimeImputer
@@ -175,10 +173,7 @@ bool_pipeline = Pipeline(
     steps=[
         ("cast_type", FunctionTransformer(lambda df: df.astype(object))),
         ("imputers", ColumnTransformer(bool_imputers, remainder="passthrough")),
-        (
-            "onehot",
-            SklearnOneHotEncoder(sparse=False, handle_unknown="ignore", drop="first"),
-        ),
+        ("onehot", SklearnOneHotEncoder(sparse=False, handle_unknown="ignore", drop="first")),
     ]
 )
 
@@ -235,9 +230,7 @@ print(categorical_cols)
 categorical_cols_value_counts = {}
 
 for col in categorical_cols:
-    value_counts = spark.sql(
-        f"SELECT COUNT(DISTINCT {col}) as count FROM lr_nhl_demo.dev.gold_model_stats_delta_v2"
-    ).toPandas()
+    value_counts = spark.sql(f"SELECT COUNT(DISTINCT {col}) as count FROM lr_nhl_demo.dev.gold_model_stats_delta_v2").toPandas()
     categorical_cols_value_counts[col] = value_counts
 
 categorical_cols_value_counts
@@ -253,7 +246,6 @@ from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.preprocessing import OrdinalEncoder
 from sklearn.pipeline import Pipeline
 
-
 class CategoricalIndexer(BaseEstimator, TransformerMixin):
     def __init__(self, categorical_cols):
         self.categorical_cols = categorical_cols
@@ -261,9 +253,7 @@ class CategoricalIndexer(BaseEstimator, TransformerMixin):
 
     def fit(self, X, y=None):
         for col in self.categorical_cols:
-            encoder = OrdinalEncoder(
-                handle_unknown="use_encoded_value", unknown_value=-1
-            )
+            encoder = OrdinalEncoder(handle_unknown='use_encoded_value', unknown_value=-1)
             encoder.fit(X[[col]])
             self.encoders[col] = encoder
         return self
@@ -275,11 +265,11 @@ class CategoricalIndexer(BaseEstimator, TransformerMixin):
         return X_copy
 
 
-team_categorical_cols = ["playerTeam", "previous_opposingTeam", "opposingTeam"]
+team_categorical_cols = ['playerTeam', 'previous_opposingTeam', 'opposingTeam']
 
 indexer_pipeline = Pipeline(
     steps=[
-        ("categorical_indexer", CategoricalIndexer(team_categorical_cols)),
+        ('categorical_indexer', CategoricalIndexer(team_categorical_cols)),
     ]
 )
 
@@ -301,15 +291,15 @@ one_hot_pipeline = Pipeline(
     ]
 )
 
-categorical_one_hot_transformers = [("onehot", one_hot_pipeline, ["position"])]
+categorical_one_hot_transformers = [("onehot", one_hot_pipeline, ['position'])]
 
 # COMMAND ----------
 
 from sklearn.compose import ColumnTransformer
 
 transformers = (
-    # date_transformers
-    bool_transformers
+    date_transformers
+    + bool_transformers
     + numerical_transformers
     + indexer_transformers
     + categorical_one_hot_transformers
@@ -330,7 +320,7 @@ from sklearn.feature_selection import SelectKBest, mutual_info_classif
 
 mi_pipeline = Pipeline(
     steps=[
-        ("mi_selector", SelectKBest(score_func=mutual_info_classif, k=75)),
+        ("mi_selector", SelectKBest(score_func=mutual_info_classif, k=100)),
     ]
 )
 
@@ -356,7 +346,6 @@ import pandas as pd
 from sklearn.pipeline import Pipeline, FeatureUnion
 from sklearn.ensemble import RandomForestRegressor
 
-
 class FeatureImportanceSelector(BaseEstimator, TransformerMixin):
     def __init__(self, model, n_features):
         self.model = model
@@ -364,10 +353,10 @@ class FeatureImportanceSelector(BaseEstimator, TransformerMixin):
         self.feature_names_in_ = None
 
     def fit(self, X, y=None):
-        self.feature_names_in_ = X.columns if hasattr(X, "columns") else None
+        self.feature_names_in_ = X.columns if hasattr(X, 'columns') else None
         self.model.fit(X, y)
         self.importances_ = self.model.feature_importances_
-        self.indices_ = np.argsort(self.importances_)[-self.n_features :]
+        self.indices_ = np.argsort(self.importances_)[-self.n_features:]
         return self
 
     def transform(self, X):
@@ -379,25 +368,24 @@ class FeatureImportanceSelector(BaseEstimator, TransformerMixin):
         if self.feature_names_in_ is not None:
             return np.array(self.feature_names_in_)[self.indices_]
         else:
-            return np.array([f"feature_{i}" for i in range(len(self.indices_))])
+            return np.array([f'feature_{i}' for i in range(len(self.indices_))])
 
-
+        
 rf_pipeline = Pipeline(
     steps=[
-        (
-            "rf_selector",
-            FeatureImportanceSelector(RandomForestRegressor(), n_features=200),
-        ),
+        ("rf_selector", RandomForestRegressor(n_estimators=5, random_state=42)),
     ]
 )
 
 # COMMAND ----------
 
+from sklearn.feature_selection import SelectFromModel, SelectKBest, mutual_info_classif, RFE
+
 feature_union = FeatureUnion(
     [
-        ("mi_pipeline", mi_pipeline),
-        ("rfe_pipeline", rfe_pipeline),
-        ("rf_pipeline", rf_pipeline),
+        ("mi_pipeline", SelectFromModel(mi_pipeline)),
+        ("rfe_pipeline", SelectFromModel(rfe_pipeline)),
+        ("rf_pipeline", SelectFromModel(rf_pipeline)),
     ]
 )
 
@@ -411,7 +399,7 @@ feature_union = FeatureUnion(
 # MAGIC - Test (20% of the dataset used to report the true performance of the model on an unseen dataset)
 # MAGIC
 # MAGIC `_automl_split_col_0000` contains the information of which set a given row belongs to.
-# MAGIC We use this column to split the dataset into the above 3 sets.
+# MAGIC We use this column to split the dataset into the above 3 sets. 
 # MAGIC The column should not be used for training so it is dropped after split is done.
 # MAGIC
 # MAGIC Given that `gameDate` is provided as the `time_col`, the data is split based on time order,
@@ -420,58 +408,54 @@ feature_union = FeatureUnion(
 # COMMAND ----------
 
 # DBTITLE 1,random split
-from sklearn.model_selection import train_test_split
+# from sklearn.model_selection import train_test_split
 
-# Convert df_loaded to Pandas DataFrame
-df_loaded_pd = df_loaded.toPandas()
+# # Convert df_loaded to Pandas DataFrame
+# df_loaded_pd = df_loaded.toPandas()
 
-# Separate target column from features
-X = df_loaded_pd.drop([target_col], axis=1)
-y = df_loaded_pd[target_col]
+# # Separate target column from features
+# X = df_loaded_pd.drop([target_col], axis=1)
+# y = df_loaded_pd[target_col]
 
-# Split the data into train and test datasets
-X_train, X_test, y_train, y_test = train_test_split(
-    X, y, test_size=0.2, random_state=42
-)
+# # Split the data into train and test datasets
+# X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
 
-# Split the train dataset into train and validation datasets
-X_train, X_val, y_train, y_val = train_test_split(
-    X_train, y_train, test_size=0.25, random_state=42
-)
+# # Split the train dataset into train and validation datasets
+# X_train, X_val, y_train, y_val = train_test_split(X_train, y_train, test_size=0.25, random_state=42)
 
 # COMMAND ----------
 
 # DBTITLE 1,time col split
-# # Convert df_loaded to Pandas DataFrame
-# df_loaded_pd = df_loaded.toPandas()
+# Convert df_loaded to Pandas DataFrame
+df_loaded_pd = df_loaded.toPandas()
 
-# # Sort the DataFrame based on the date column
-# df_loaded_pd = df_loaded_pd.sort_values(time_col)
+# Sort the DataFrame based on the date column
+df_loaded_pd = df_loaded_pd.sort_values(time_col)
 
-# # Determine the indices to split the DataFrame
-# train_size = int(0.6 * len(df_loaded_pd))
-# val_size = int(0.2 * len(df_loaded_pd))
+# Determine the indices to split the DataFrame
+train_size = int(0.6 * len(df_loaded_pd))
+val_size = int(0.2 * len(df_loaded_pd))
 
-# train_indices = list(range(train_size))
-# val_indices = list(range(train_size, train_size + val_size))
-# test_indices = list(range(train_size + val_size, len(df_loaded_pd)))
+train_indices = list(range(train_size))
+val_indices = list(range(train_size, train_size + val_size))
+test_indices = list(range(train_size + val_size, len(df_loaded_pd)))
 
-# # Split the DataFrame into training, validation, and test sets
-# split_train_df = df_loaded_pd.iloc[train_indices]
-# split_val_df = df_loaded_pd.iloc[val_indices]
-# split_test_df = df_loaded_pd.iloc[test_indices]
+# Split the DataFrame into training, validation, and test sets
+split_train_df = df_loaded_pd.iloc[train_indices]
+split_val_df = df_loaded_pd.iloc[val_indices]
+split_test_df = df_loaded_pd.iloc[test_indices]
 
 # COMMAND ----------
 
-# # Separate target column from features and drop _automl_split_col_0000
-# X_train = split_train_df.drop([target_col], axis=1)
-# y_train = split_train_df[target_col]
+# Separate target column from features and drop _automl_split_col_0000
+X_train = split_train_df.drop([target_col], axis=1)
+y_train = split_train_df[target_col]
 
-# X_val = split_val_df.drop([target_col], axis=1)
-# y_val = split_val_df[target_col]
+X_val = split_val_df.drop([target_col], axis=1)
+y_val = split_val_df[target_col]
 
-# X_test = split_test_df.drop([target_col], axis=1)
-# y_test = split_test_df[target_col]
+X_test = split_test_df.drop([target_col], axis=1)
+y_test = split_test_df[target_col]
 
 # COMMAND ----------
 
@@ -481,12 +465,6 @@ X_train, X_val, y_train, y_val = train_test_split(
 # COMMAND ----------
 
 from sklearn.ensemble import RandomForestRegressor
-from sklearn.feature_selection import (
-    SelectFromModel,
-    SelectKBest,
-    mutual_info_classif,
-    RFE,
-)
 from sklearn.linear_model import LogisticRegression
 from databricks.automl_runtime.sklearn.column_selector import ColumnSelector
 
@@ -494,37 +472,15 @@ import lightgbm
 from lightgbm import LGBMRegressor
 
 from sklearn import set_config
-
 set_config(transform_output="pandas")
-
-# Now create the FeatureUnion with fitted pipelines
-# feature_union = FeatureUnion(
-#     [
-# ("mi_pipeline", ColumnSelector(columns=mi_pipeline.named_steps['selector'].get_feature_names_out())),
-# ("rfe_pipeline", ColumnSelector(columns=rfe_pipeline.named_steps['selector'].get_feature_names_out())),
-# ("rf_pipeline", ColumnSelector(columns=SelectFromModel(RandomForestRegressor(n_estimators=5, random_state=42)).get_feature_names_out())),
-#     ]
-# )
-
-# features_pipeline = Pipeline(
-#     [
-#         ("column_selector", col_selector),
-#         ("preprocessor", preprocessor),
-#         ("rf", RandomForestRegressor(n_estimators=5, random_state=42)),
-#         ("feature_selector", SelectFromModel(rf, prefit=True)),
-#         ("regressor", LGBMRegressor()),
-#     ]
-# )
 
 # Create the full pipeline
 features_pipeline = Pipeline(
     [
         ("column_selector", col_selector),
         ("preprocessor", preprocessor),
-        (
-            "feature_selector",
-            SelectFromModel(RandomForestRegressor(n_estimators=5, random_state=42)),
-        ),
+        # ("feature_selector", SelectFromModel(RandomForestRegressor(n_estimators=5, random_state=42))),
+        ("feature_selector", feature_union),
     ]
 )
 
@@ -549,7 +505,6 @@ set_config(transform_output="pandas")
 
 features_pipeline.fit(X_train, y_train)
 
-
 # COMMAND ----------
 
 processed_X_train = features_pipeline.transform(X_train)
@@ -562,7 +517,7 @@ processed_X_train
 # COMMAND ----------
 
 # Get selected columns
-selected_columns = features_pipeline.named_steps["feature_selector"].get_support()
+selected_columns = features_pipeline.named_steps['feature_selector'].get_support()
 print("Selected feature names count:", len(selected_columns))
 
 # TO DO: convert X_train below to be the preprocessed dataset
@@ -593,215 +548,36 @@ help(LGBMRegressor)
 
 # COMMAND ----------
 
-# DBTITLE 1,log preprocessing pipeline
-import mlflow
-import os
-from mlflow.models import Model, infer_signature, ModelSignature
-from mlflow.pyfunc import PythonModel
-import sklearn
-from sklearn import set_config
-from sklearn.pipeline import Pipeline
-from sklearn.ensemble import RandomForestRegressor
-from sklearn.feature_selection import SelectFromModel
-import joblib
-
-
-class PreprocessModel(PythonModel):
-    def __init__(self, pipeline):
-        self.pipeline = pipeline
-
-    def fit(self, X, y):
-        self.pipeline.fit(X, y)
-
-    def predict(self, context, model_input):
-        return self.pipeline.transform(model_input)
-
-
-# Create the preprocessing pipeline
-preprocess_pipeline = Pipeline(
-    [
-        ("column_selector", col_selector),
-        ("preprocessor", preprocessor),
-        (
-            "feature_selector",
-            SelectFromModel(RandomForestRegressor(n_estimators=5, random_state=42)),
-        ),
-    ]
-)
-
-# Create an instance of the custom PythonModel
-pyfunc_preprocess_model = PreprocessModel(preprocess_pipeline)
-
-# Fit the model
-pyfunc_preprocess_model.fit(X_train, y_train)
-
-# Get the current working directory
-cwd = os.getcwd()
-
-# Specify the file name
-file_name = 'preprocess_model'
-
-# Create the full file path by joining the current working directory with the file name
-path = os.path.join(cwd, file_name)
-
-# Check if the directory exists
-if os.path.exists(path):
-    # Check if the path is a directory
-    if os.path.isdir(path):
-        # Remove the directory and its contents
-        shutil.rmtree(path)
-        print(f"The directory '{path}' and its contents have been deleted.")
-    else:
-        print(f"The path '{path}' is not a directory.")
-else:
-    print(f"The directory '{path}' does not exist.")
-
-# Save the model using MLflow
-with mlflow.start_run() as run:
-    mlflow.pyfunc.save_model(
-        path="preprocess_model",
-        python_model=pyfunc_preprocess_model,
-        input_example=X_train.iloc[:5],
-        signature=infer_signature(X_train, preprocess_pipeline.transform(X_train)),
-    )
-
-    # Log the model
-    mlflow.pyfunc.log_model(
-        artifact_path="preprocess_model",
-        python_model=pyfunc_preprocess_model,
-        input_example=X_train.iloc[:5],
-        signature=infer_signature(X_train, preprocess_pipeline.transform(X_train)),
-    )
-
-# Load the saved model
-loaded_model = mlflow.pyfunc.load_model("preprocess_model")
-
-# Transform datasets using the loaded model
-X_val_processed = loaded_model.predict(X_val)
-X_train_processed = loaded_model.predict(X_train)
-X_test_processed = loaded_model.predict(X_test)
-
-# COMMAND ----------
-
-import mlflow
-import os
-import shutil
-import tempfile
-import yaml
-
-preprocess_run = mlflow.last_active_run()
-run_id = preprocess_run.info.run_id
-
-# Set up a local dir for downloading the artifacts.
-tmp_dir = tempfile.mkdtemp()
-
-client = mlflow.tracking.MlflowClient()
-
-# Fix conda.yaml
-conda_file_path = mlflow.artifacts.download_artifacts(
-    artifact_uri=f"runs:/{run_id}/preprocess_model/conda.yaml", dst_path=tmp_dir
-)
-with open(conda_file_path) as f:
-    conda_libs = yaml.load(f, Loader=yaml.FullLoader)
-pandas_lib_exists = any(
-    [lib.startswith("pandas==") for lib in conda_libs["dependencies"][-1]["pip"]]
-)
-if not pandas_lib_exists:
-    print("Adding pandas dependency to conda.yaml")
-    conda_libs["dependencies"][-1]["pip"].append(f"pandas=={pd.__version__}")
-
-    with open(f"{tmp_dir}/conda.yaml", "w") as f:
-        f.write(yaml.dump(conda_libs))
-    client.log_artifact(
-        run_id=run_id, local_path=conda_file_path, artifact_path="preprocess_model"
-    )
-
-# Fix requirements.txt
-venv_file_path = mlflow.artifacts.download_artifacts(
-    artifact_uri=f"runs:/{run_id}/preprocess_model/requirements.txt", dst_path=tmp_dir
-)
-with open(venv_file_path) as f:
-    venv_libs = f.readlines()
-venv_libs = [lib.strip() for lib in venv_libs]
-pandas_lib_exists = any([lib.startswith("pandas==") for lib in venv_libs])
-if not pandas_lib_exists:
-    print("Adding pandas dependency to requirements.txt")
-    venv_libs.append(f"pandas=={pd.__version__}")
-
-    with open(f"{tmp_dir}/requirements.txt", "w") as f:
-        f.write("\n".join(venv_libs))
-    client.log_artifact(
-        run_id=run_id, local_path=venv_file_path, artifact_path="preprocess_model"
-    )
-
-shutil.rmtree(tmp_dir)
-
-# COMMAND ----------
-
-# Get the run ID and model URI
-preprocess_model_uri = f"runs:/{preprocess_run.info.run_id}/preprocess_model"
-
-# Register the model
-mlflow.register_model(preprocess_model_uri, "lr_nhl_demo.dev.preprocess_model")
-
-# COMMAND ----------
-
-# Set as Champion based on max version
-client = mlflow.tracking.MlflowClient()
-model_version_infos = client.search_model_versions("name = 'lr_nhl_demo.dev.preprocess_model'")
-new_model_version = max([model_version_info.version for model_version_info in model_version_infos])
-client.set_registered_model_alias("lr_nhl_demo.dev.preprocess_model", "champion", new_model_version)
-
-pp_champion_version = client.get_model_version_by_alias(
-    "lr_nhl_demo.dev.preprocess_model", "champion"
-)
-
-preprocess_model_name = pp_champion_version.name
-preprocess_model_version = pp_champion_version.version
-
-preprocess_model_uri = f"models:/{preprocess_model_name}/{preprocess_model_version}"
-preprocess_model = mlflow.pyfunc.load_model(model_uri=preprocess_model_uri)
-
-# COMMAND ----------
-
-X_train_processed = preprocess_model.predict(X_train)
-
-# COMMAND ----------
-
-X_train_processed
-
-# COMMAND ----------
-
-X_train_processed_UC = mlflow.data.load_delta(table_name="lr_nhl_demo.dev.X_train_processed", version="0")
-X_train_processed_UC_PD = X_train_processed_UC.df.toPandas()
-
-X_train_processed_UC_PD
-
-# COMMAND ----------
-
-X_train_processed_UC_PD
-
-# COMMAND ----------
-
 import mlflow
 from mlflow.models import Model, infer_signature, ModelSignature
 from mlflow.pyfunc import PyFuncModel
 from mlflow import pyfunc
-from hyperopt import hp, tpe, fmin, STATUS_OK, SparkTrials
-import pyspark.pandas as ps
+import sklearn
+from sklearn import set_config
+from sklearn.pipeline import Pipeline
+from hyperopt import hp, tpe, fmin, STATUS_OK, Trials
 
-# Write X_train_processed to Unity Catalog for lineage tracking
-ps.from_pandas(X_train_processed).to_table("lr_nhl_demo.dev.X_train_processed", mode="overwrite")
-print("Writing X_train_processed to UC!")
 
-# Load a Unity Catalog table, train a model, and log the input table
-print("Loading X_train_processed_UC")
-X_train_processed_UC = mlflow.data.load_delta(table_name="lr_nhl_demo.dev.X_train_processed", version="0")
-X_train_processed_UC_PD = X_train_processed_UC.df.toPandas()
+# Create a separate pipeline to transform the validation dataset. This is used for early stopping.
+preprocess_pipeline = Pipeline(
+    [
+        ("column_selector", col_selector),
+        ("preprocessor", preprocessor),
+        ("feature_selector", SelectFromModel(RandomForestRegressor(n_estimators=5, random_state=42))),
+    ]
+)
+
+mlflow.sklearn.autolog(disable=True)
+preprocess_pipeline.fit(X_train, y_train)
+X_val_processed = preprocess_pipeline.transform(X_val)
+
+X_train_processed = preprocess_pipeline.transform(X_train)
+X_test_processed = preprocess_pipeline.transform(X_test)
 
 def objective(params):
-    with mlflow.start_run(experiment_id="634720160613016") as mlflow_run:
+    with mlflow.start_run(experiment_id="2824690123542843") as mlflow_run:
         lgbmr_regressor = LGBMRegressor(**params)
+
         model = Pipeline(
             [
                 ("regressor", lgbmr_regressor),
@@ -823,7 +599,6 @@ def objective(params):
             ],
             regressor__eval_set=[(X_val_processed, y_val)],
         )
-        # mlflow.log_input(X_train_processed_UC_PD, "training_data")
 
         # Log metrics for the training set
         mlflow_model = Model()
@@ -874,6 +649,12 @@ def objective(params):
             k.replace("test_", ""): v for k, v in lgbmr_test_metrics.items()
         }
 
+        # # Add these lines to print the selected features
+        # feature_names = model.named_steps['preprocessor'].get_feature_names_out()
+        # selected_features = model.named_steps['feature_selection'].get_feature_names_out()
+        # print("Selected features after pipeline:", selected_features)
+        # mlflow.log_param("selected_features", selected_features.tolist())
+
         return {
             "loss": loss,
             "status": STATUS_OK,
@@ -883,8 +664,11 @@ def objective(params):
             "run": mlflow_run,
         }
 
-
 X_train_processed
+
+# COMMAND ----------
+
+
 
 # COMMAND ----------
 
@@ -909,21 +693,18 @@ X_train_processed
 
 # COMMAND ----------
 
-from hyperopt.pyll.base import scope
-
-# Define the search space
 space = {
-    "colsample_bytree": hp.uniform("colsample_bytree", 0.5, 1.0),
-    "lambda_l1": hp.loguniform("lambda_l1", -5, 0),  # loguniform for positive values
-    "lambda_l2": hp.loguniform("lambda_l2", -5, 2),  # loguniform for positive values
-    "learning_rate": hp.loguniform("learning_rate", -5, -1),  # loguniform for small positive values
-    "max_bin": scope.int(hp.quniform("max_bin", 20, 100, 1)),  # integer values
-    "max_depth": scope.int(hp.quniform("max_depth", 3, 15, 1)),  # integer values
-    "min_child_samples": scope.int(hp.quniform("min_child_samples", 20, 200, 1)),  # integer values
-    "n_estimators": scope.int(hp.quniform("n_estimators", 100, 500, 1)),  # integer values
-    "num_leaves": scope.int(hp.quniform("num_leaves", 31, 255, 1)),  # integer values
-    "subsample": hp.uniform("subsample", 0.5, 1.0),
-    "random_state": 729986891,  # Fixed value, not part of the search space
+    "colsample_bytree": 0.6587254729574785,
+    "lambda_l1": 0.18965542130830132,
+    "lambda_l2": 1.9402892441576718,
+    "learning_rate": 0.02517876184843777,
+    "max_bin": 29,
+    "max_depth": 8,
+    "min_child_samples": 140,
+    "n_estimators": 287,
+    "num_leaves": 166,
+    "subsample": 0.7501639626721491,
+    "random_state": 729986891,
 }
 
 # COMMAND ----------
@@ -944,7 +725,7 @@ space = {
 
 # COMMAND ----------
 
-# from hyperopt import SparkTrials, fmin, tpe
+from hyperopt import SparkTrials
 trials = SparkTrials()
 
 # COMMAND ----------
@@ -954,7 +735,7 @@ fmin(
     objective,
     space=space,
     algo=tpe.suggest,
-    max_evals=25,  # Increase this when widening the hyperparameter search space.
+    max_evals=1,  # Increase this when widening the hyperparameter search space.
     trials=trials,
 )
 
@@ -1074,14 +855,12 @@ if shap_enabled:
     ).fillna(mode)
 
     # Sample some rows from the validation set to explain. Increase the sample size for more thorough results.
-    example = X_val_processed.sample(
-        n=min(100, X_val_processed.shape[0]), random_state=729986891
-    ).fillna(mode)
+    example = X_val_processed.sample(n=min(100, X_val_processed.shape[0]), random_state=729986891).fillna(
+        mode
+    )
 
     # Use Kernel SHAP to explain feature importance on the sampled rows from the validation set.
-    predict = lambda x: model.predict(
-        pd.DataFrame(x, columns=X_train_processed.columns)
-    )
+    predict = lambda x: model.predict(pd.DataFrame(x, columns=X_train_processed.columns))
     explainer = KernelExplainer(predict, train_sample, link="identity")
     shap_values = explainer.shap_values(example, l1_reg=False, nsamples=500)
     summary_plot(shap_values, example)
@@ -1124,3 +903,7 @@ if shap_enabled:
 
 # model_uri for the generated model
 print(f"runs:/{ mlflow_run.info.run_id }/model")
+
+# COMMAND ----------
+
+
