+++ nhlPredict/src/ML/model_training/Predict_SOG_v2-Manual.py
@@ -9,7 +9,6 @@
 # COMMAND ----------
 
 import mlflow
-import databricks.automl_runtime
 
 target_col = "player_Total_shotsOnGoal"
 time_col = "gameDate"
@@ -21,16 +20,13 @@ time_col = "gameDate"
 
 # COMMAND ----------
 
-import mlflow
-import os
-import uuid
 import shutil
 import pandas as pd
 
 df_loaded = spark.table("lr_nhl_demo.dev.SOG_features_v2")
 
 # Preview data
-display(df_loaded.head(5))
+display(df_loaded.limit(5))
 
 # COMMAND ----------
 
@@ -64,9 +60,16 @@ display(df_loaded.head(5))
 
 # COMMAND ----------
 
-from pyspark.sql.types import StringType
-
-cols_to_remove = ["DAY", "HOME", "AWAY", "gameId", "playerId", "shooterName", "home_or_away"]
+cols_to_remove = [
+    "DAY",
+    "HOME",
+    "AWAY",
+    "gameId",
+    "playerId",
+    "shooterName",
+    "home_or_away",
+    time_col,
+]
 
 # Identify numerical and categorical columns
 numerical_cols = [
@@ -88,8 +91,8 @@ print(numerical_cols)
 from databricks.automl_runtime.sklearn.column_selector import ColumnSelector
 
 supported_cols = categorical_cols + [col for col in numerical_cols if col != target_col]
-supported_cols = list(set(supported_cols + [time_col]) - set(cols_to_remove))
-# supported_cols = list(set(supported_cols) - set(cols_to_remove))
+# supported_cols = list(set(supported_cols + [time_col]) - set(cols_to_remove))
+supported_cols = list(set(supported_cols) - set(cols_to_remove))
 
 col_selector = ColumnSelector(supported_cols)
 
@@ -129,7 +132,6 @@ target_col in numerical_cols
 
 # COMMAND ----------
 
-from pandas import Timestamp
 from sklearn.pipeline import Pipeline
 
 from databricks.automl_runtime.sklearn import DatetimeImputer
@@ -173,7 +175,10 @@ bool_pipeline = Pipeline(
     steps=[
         ("cast_type", FunctionTransformer(lambda df: df.astype(object))),
         ("imputers", ColumnTransformer(bool_imputers, remainder="passthrough")),
-        ("onehot", SklearnOneHotEncoder(sparse=False, handle_unknown="ignore", drop="first")),
+        (
+            "onehot",
+            SklearnOneHotEncoder(sparse=False, handle_unknown="ignore", drop="first"),
+        ),
     ]
 )
 
@@ -230,7 +235,9 @@ print(categorical_cols)
 categorical_cols_value_counts = {}
 
 for col in categorical_cols:
-    value_counts = spark.sql(f"SELECT COUNT(DISTINCT {col}) as count FROM lr_nhl_demo.dev.gold_model_stats_delta_v2").toPandas()
+    value_counts = spark.sql(
+        f"SELECT COUNT(DISTINCT {col}) as count FROM lr_nhl_demo.dev.gold_model_stats_delta_v2"
+    ).toPandas()
     categorical_cols_value_counts[col] = value_counts
 
 categorical_cols_value_counts
@@ -246,6 +253,7 @@ from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.preprocessing import OrdinalEncoder
 from sklearn.pipeline import Pipeline
 
+
 class CategoricalIndexer(BaseEstimator, TransformerMixin):
     def __init__(self, categorical_cols):
         self.categorical_cols = categorical_cols
@@ -253,7 +261,9 @@ class CategoricalIndexer(BaseEstimator, TransformerMixin):
 
     def fit(self, X, y=None):
         for col in self.categorical_cols:
-            encoder = OrdinalEncoder(handle_unknown='use_encoded_value', unknown_value=-1)
+            encoder = OrdinalEncoder(
+                handle_unknown="use_encoded_value", unknown_value=-1
+            )
             encoder.fit(X[[col]])
             self.encoders[col] = encoder
         return self
@@ -265,11 +275,11 @@ class CategoricalIndexer(BaseEstimator, TransformerMixin):
         return X_copy
 
 
-team_categorical_cols = ['playerTeam', 'previous_opposingTeam', 'opposingTeam']
+team_categorical_cols = ["playerTeam", "previous_opposingTeam", "opposingTeam"]
 
 indexer_pipeline = Pipeline(
     steps=[
-        ('categorical_indexer', CategoricalIndexer(team_categorical_cols)),
+        ("categorical_indexer", CategoricalIndexer(team_categorical_cols)),
     ]
 )
 
@@ -291,15 +301,15 @@ one_hot_pipeline = Pipeline(
     ]
 )
 
-categorical_one_hot_transformers = [("onehot", one_hot_pipeline, ['position'])]
+categorical_one_hot_transformers = [("onehot", one_hot_pipeline, ["position"])]
 
 # COMMAND ----------
 
 from sklearn.compose import ColumnTransformer
 
 transformers = (
-    date_transformers
-    + bool_transformers
+    # date_transformers
+    bool_transformers
     + numerical_transformers
     + indexer_transformers
     + categorical_one_hot_transformers
@@ -320,7 +330,7 @@ from sklearn.feature_selection import SelectKBest, mutual_info_classif
 
 mi_pipeline = Pipeline(
     steps=[
-        ("mi_selector", SelectKBest(score_func=mutual_info_classif, k=100)),
+        ("mi_selector", SelectKBest(score_func=mutual_info_classif, k=75)),
     ]
 )
 
@@ -346,6 +356,7 @@ import pandas as pd
 from sklearn.pipeline import Pipeline, FeatureUnion
 from sklearn.ensemble import RandomForestRegressor
 
+
 class FeatureImportanceSelector(BaseEstimator, TransformerMixin):
     def __init__(self, model, n_features):
         self.model = model
@@ -353,10 +364,10 @@ class FeatureImportanceSelector(BaseEstimator, TransformerMixin):
         self.feature_names_in_ = None
 
     def fit(self, X, y=None):
-        self.feature_names_in_ = X.columns if hasattr(X, 'columns') else None
+        self.feature_names_in_ = X.columns if hasattr(X, "columns") else None
         self.model.fit(X, y)
         self.importances_ = self.model.feature_importances_
-        self.indices_ = np.argsort(self.importances_)[-self.n_features:]
+        self.indices_ = np.argsort(self.importances_)[-self.n_features :]
         return self
 
     def transform(self, X):
@@ -368,24 +379,25 @@ class FeatureImportanceSelector(BaseEstimator, TransformerMixin):
         if self.feature_names_in_ is not None:
             return np.array(self.feature_names_in_)[self.indices_]
         else:
-            return np.array([f'feature_{i}' for i in range(len(self.indices_))])
+            return np.array([f"feature_{i}" for i in range(len(self.indices_))])
 
-        
+
 rf_pipeline = Pipeline(
     steps=[
-        ("rf_selector", RandomForestRegressor(n_estimators=5, random_state=42)),
+        (
+            "rf_selector",
+            FeatureImportanceSelector(RandomForestRegressor(), n_features=200),
+        ),
     ]
 )
 
 # COMMAND ----------
 
-from sklearn.feature_selection import SelectFromModel, SelectKBest, mutual_info_classif, RFE
-
 feature_union = FeatureUnion(
     [
-        ("mi_pipeline", SelectFromModel(mi_pipeline)),
-        ("rfe_pipeline", SelectFromModel(rfe_pipeline)),
-        ("rf_pipeline", SelectFromModel(rf_pipeline)),
+        ("mi_pipeline", mi_pipeline),
+        ("rfe_pipeline", rfe_pipeline),
+        ("rf_pipeline", rf_pipeline),
     ]
 )
 
@@ -399,7 +411,7 @@ feature_union = FeatureUnion(
 # MAGIC - Test (20% of the dataset used to report the true performance of the model on an unseen dataset)
 # MAGIC
 # MAGIC `_automl_split_col_0000` contains the information of which set a given row belongs to.
-# MAGIC We use this column to split the dataset into the above 3 sets. 
+# MAGIC We use this column to split the dataset into the above 3 sets.
 # MAGIC The column should not be used for training so it is dropped after split is done.
 # MAGIC
 # MAGIC Given that `gameDate` is provided as the `time_col`, the data is split based on time order,
@@ -408,54 +420,58 @@ feature_union = FeatureUnion(
 # COMMAND ----------
 
 # DBTITLE 1,random split
-# from sklearn.model_selection import train_test_split
+from sklearn.model_selection import train_test_split
 
-# # Convert df_loaded to Pandas DataFrame
-# df_loaded_pd = df_loaded.toPandas()
+# Convert df_loaded to Pandas DataFrame
+df_loaded_pd = df_loaded.toPandas()
 
-# # Separate target column from features
-# X = df_loaded_pd.drop([target_col], axis=1)
-# y = df_loaded_pd[target_col]
+# Separate target column from features
+X = df_loaded_pd.drop([target_col], axis=1)
+y = df_loaded_pd[target_col]
 
-# # Split the data into train and test datasets
-# X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
+# Split the data into train and test datasets
+X_train, X_test, y_train, y_test = train_test_split(
+    X, y, test_size=0.2, random_state=42
+)
 
-# # Split the train dataset into train and validation datasets
-# X_train, X_val, y_train, y_val = train_test_split(X_train, y_train, test_size=0.25, random_state=42)
+# Split the train dataset into train and validation datasets
+X_train, X_val, y_train, y_val = train_test_split(
+    X_train, y_train, test_size=0.25, random_state=42
+)
 
 # COMMAND ----------
 
 # DBTITLE 1,time col split
-# Convert df_loaded to Pandas DataFrame
-df_loaded_pd = df_loaded.toPandas()
+# # Convert df_loaded to Pandas DataFrame
+# df_loaded_pd = df_loaded.toPandas()
 
-# Sort the DataFrame based on the date column
-df_loaded_pd = df_loaded_pd.sort_values(time_col)
+# # Sort the DataFrame based on the date column
+# df_loaded_pd = df_loaded_pd.sort_values(time_col)
 
-# Determine the indices to split the DataFrame
-train_size = int(0.6 * len(df_loaded_pd))
-val_size = int(0.2 * len(df_loaded_pd))
+# # Determine the indices to split the DataFrame
+# train_size = int(0.6 * len(df_loaded_pd))
+# val_size = int(0.2 * len(df_loaded_pd))
 
-train_indices = list(range(train_size))
-val_indices = list(range(train_size, train_size + val_size))
-test_indices = list(range(train_size + val_size, len(df_loaded_pd)))
+# train_indices = list(range(train_size))
+# val_indices = list(range(train_size, train_size + val_size))
+# test_indices = list(range(train_size + val_size, len(df_loaded_pd)))
 
-# Split the DataFrame into training, validation, and test sets
-split_train_df = df_loaded_pd.iloc[train_indices]
-split_val_df = df_loaded_pd.iloc[val_indices]
-split_test_df = df_loaded_pd.iloc[test_indices]
+# # Split the DataFrame into training, validation, and test sets
+# split_train_df = df_loaded_pd.iloc[train_indices]
+# split_val_df = df_loaded_pd.iloc[val_indices]
+# split_test_df = df_loaded_pd.iloc[test_indices]
 
 # COMMAND ----------
 
-# Separate target column from features and drop _automl_split_col_0000
-X_train = split_train_df.drop([target_col], axis=1)
-y_train = split_train_df[target_col]
+# # Separate target column from features and drop _automl_split_col_0000
+# X_train = split_train_df.drop([target_col], axis=1)
+# y_train = split_train_df[target_col]
 
-X_val = split_val_df.drop([target_col], axis=1)
-y_val = split_val_df[target_col]
+# X_val = split_val_df.drop([target_col], axis=1)
+# y_val = split_val_df[target_col]
 
-X_test = split_test_df.drop([target_col], axis=1)
-y_test = split_test_df[target_col]
+# X_test = split_test_df.drop([target_col], axis=1)
+# y_test = split_test_df[target_col]
 
 # COMMAND ----------
 
@@ -465,6 +481,12 @@ y_test = split_test_df[target_col]
 # COMMAND ----------
 
 from sklearn.ensemble import RandomForestRegressor
+from sklearn.feature_selection import (
+    SelectFromModel,
+    SelectKBest,
+    mutual_info_classif,
+    RFE,
+)
 from sklearn.linear_model import LogisticRegression
 from databricks.automl_runtime.sklearn.column_selector import ColumnSelector
 
@@ -472,15 +494,37 @@ import lightgbm
 from lightgbm import LGBMRegressor
 
 from sklearn import set_config
+
 set_config(transform_output="pandas")
+
+# Now create the FeatureUnion with fitted pipelines
+# feature_union = FeatureUnion(
+#     [
+# ("mi_pipeline", ColumnSelector(columns=mi_pipeline.named_steps['selector'].get_feature_names_out())),
+# ("rfe_pipeline", ColumnSelector(columns=rfe_pipeline.named_steps['selector'].get_feature_names_out())),
+# ("rf_pipeline", ColumnSelector(columns=SelectFromModel(RandomForestRegressor(n_estimators=5, random_state=42)).get_feature_names_out())),
+#     ]
+# )
+
+# features_pipeline = Pipeline(
+#     [
+#         ("column_selector", col_selector),
+#         ("preprocessor", preprocessor),
+#         ("rf", RandomForestRegressor(n_estimators=5, random_state=42)),
+#         ("feature_selector", SelectFromModel(rf, prefit=True)),
+#         ("regressor", LGBMRegressor()),
+#     ]
+# )
 
 # Create the full pipeline
 features_pipeline = Pipeline(
     [
         ("column_selector", col_selector),
         ("preprocessor", preprocessor),
-        # ("feature_selector", SelectFromModel(RandomForestRegressor(n_estimators=5, random_state=42))),
-        ("feature_selector", feature_union),
+        (
+            "feature_selector",
+            SelectFromModel(RandomForestRegressor(n_estimators=5, random_state=42)),
+        ),
     ]
 )
 
@@ -505,6 +549,7 @@ set_config(transform_output="pandas")
 
 features_pipeline.fit(X_train, y_train)
 
+
 # COMMAND ----------
 
 processed_X_train = features_pipeline.transform(X_train)
@@ -517,7 +562,7 @@ processed_X_train
 # COMMAND ----------
 
 # Get selected columns
-selected_columns = features_pipeline.named_steps['feature_selector'].get_support()
+selected_columns = features_pipeline.named_steps["feature_selector"].get_support()
 print("Selected feature names count:", len(selected_columns))
 
 # TO DO: convert X_train below to be the preprocessed dataset
@@ -548,39 +593,242 @@ help(LGBMRegressor)
 
 # COMMAND ----------
 
+# DBTITLE 1,log preprocessing pipeline
 import mlflow
+import os
 from mlflow.models import Model, infer_signature, ModelSignature
-from mlflow.pyfunc import PyFuncModel
-from mlflow import pyfunc
+from mlflow.pyfunc import PythonModel
 import sklearn
 from sklearn import set_config
 from sklearn.pipeline import Pipeline
-from hyperopt import hp, tpe, fmin, STATUS_OK, Trials
+from sklearn.ensemble import RandomForestRegressor
+from sklearn.feature_selection import SelectFromModel
+import joblib
 
 
-# Create a separate pipeline to transform the validation dataset. This is used for early stopping.
+class PreprocessModel(PythonModel):
+    def __init__(self, pipeline):
+        self.pipeline = pipeline
+
+    def fit(self, X, y):
+        self.pipeline.fit(X, y)
+
+    def predict(self, context, model_input):
+        return self.pipeline.transform(model_input)
+
+
+# Create the preprocessing pipeline
 preprocess_pipeline = Pipeline(
     [
         ("column_selector", col_selector),
         ("preprocessor", preprocessor),
-        ("feature_selector", SelectFromModel(RandomForestRegressor(n_estimators=5, random_state=42))),
+        (
+            "feature_selector",
+            SelectFromModel(RandomForestRegressor(n_estimators=5, random_state=42)),
+        ),
     ]
 )
 
-mlflow.sklearn.autolog(disable=True)
-preprocess_pipeline.fit(X_train, y_train)
-X_val_processed = preprocess_pipeline.transform(X_val)
+# Create an instance of the custom PythonModel
+pyfunc_preprocess_model = PreprocessModel(preprocess_pipeline)
 
-X_train_processed = preprocess_pipeline.transform(X_train)
-X_test_processed = preprocess_pipeline.transform(X_test)
+# Fit the model
+pyfunc_preprocess_model.fit(X_train, y_train)
 
+# Get the current working directory
+cwd = os.getcwd()
+
+# Specify the file name
+file_name = 'preprocess_model'
+
+# Create the full file path by joining the current working directory with the file name
+path = os.path.join(cwd, file_name)
+
+# Check if the directory exists
+if os.path.exists(path):
+    # Check if the path is a directory
+    if os.path.isdir(path):
+        # Remove the directory and its contents
+        shutil.rmtree(path)
+        print(f"The directory '{path}' and its contents have been deleted.")
+    else:
+        print(f"The path '{path}' is not a directory.")
+else:
+    print(f"The directory '{path}' does not exist.")
+
+# Save the model using MLflow
+with mlflow.start_run() as run:
+    mlflow.pyfunc.save_model(
+        path="preprocess_model",
+        python_model=pyfunc_preprocess_model,
+        input_example=X_train.iloc[:5],
+        signature=infer_signature(X_train, preprocess_pipeline.transform(X_train)),
+    )
+
+    # Log the model
+    mlflow.pyfunc.log_model(
+        artifact_path="preprocess_model",
+        python_model=pyfunc_preprocess_model,
+        input_example=X_train.iloc[:5],
+        signature=infer_signature(X_train, preprocess_pipeline.transform(X_train)),
+    )
+
+# Load the saved model
+loaded_model = mlflow.pyfunc.load_model("preprocess_model")
+
+# Transform datasets using the loaded model
+X_val_processed = loaded_model.predict(X_val)
+X_train_processed = loaded_model.predict(X_train)
+X_test_processed = loaded_model.predict(X_test)
+
+# COMMAND ----------
+
+import mlflow
+import os
+import shutil
+import tempfile
+import yaml
+
+preprocess_run = mlflow.last_active_run()
+run_id = preprocess_run.info.run_id
+
+# Set up a local dir for downloading the artifacts.
+tmp_dir = tempfile.mkdtemp()
+
+client = mlflow.tracking.MlflowClient()
+
+# Fix conda.yaml
+conda_file_path = mlflow.artifacts.download_artifacts(
+    artifact_uri=f"runs:/{run_id}/preprocess_model/conda.yaml", dst_path=tmp_dir
+)
+with open(conda_file_path) as f:
+    conda_libs = yaml.load(f, Loader=yaml.FullLoader)
+pandas_lib_exists = any(
+    [lib.startswith("pandas==") for lib in conda_libs["dependencies"][-1]["pip"]]
+)
+if not pandas_lib_exists:
+    print("Adding pandas dependency to conda.yaml")
+    conda_libs["dependencies"][-1]["pip"].append(f"pandas=={pd.__version__}")
+
+    with open(f"{tmp_dir}/conda.yaml", "w") as f:
+        f.write(yaml.dump(conda_libs))
+    client.log_artifact(
+        run_id=run_id, local_path=conda_file_path, artifact_path="preprocess_model"
+    )
+
+# Fix requirements.txt
+venv_file_path = mlflow.artifacts.download_artifacts(
+    artifact_uri=f"runs:/{run_id}/preprocess_model/requirements.txt", dst_path=tmp_dir
+)
+with open(venv_file_path) as f:
+    venv_libs = f.readlines()
+venv_libs = [lib.strip() for lib in venv_libs]
+pandas_lib_exists = any([lib.startswith("pandas==") for lib in venv_libs])
+if not pandas_lib_exists:
+    print("Adding pandas dependency to requirements.txt")
+    venv_libs.append(f"pandas=={pd.__version__}")
+
+    with open(f"{tmp_dir}/requirements.txt", "w") as f:
+        f.write("\n".join(venv_libs))
+    client.log_artifact(
+        run_id=run_id, local_path=venv_file_path, artifact_path="preprocess_model"
+    )
+
+shutil.rmtree(tmp_dir)
+
+# COMMAND ----------
+
+# Get the run ID and model URI
+preprocess_model_uri = f"runs:/{preprocess_run.info.run_id}/preprocess_model"
+
+# Register the model
+mlflow.register_model(preprocess_model_uri, "lr_nhl_demo.dev.preprocess_model")
+
+# COMMAND ----------
+
+# Set as Champion based on max version
+client = mlflow.tracking.MlflowClient()
+model_version_infos = client.search_model_versions("name = 'lr_nhl_demo.dev.preprocess_model'")
+new_model_version = max([model_version_info.version for model_version_info in model_version_infos])
+client.set_registered_model_alias("lr_nhl_demo.dev.preprocess_model", "champion", new_model_version)
+
+pp_champion_version = client.get_model_version_by_alias(
+    "lr_nhl_demo.dev.preprocess_model", "champion"
+)
+
+preprocess_model_name = pp_champion_version.name
+preprocess_model_version = pp_champion_version.version
+
+preprocess_model_uri = f"models:/{preprocess_model_name}/{preprocess_model_version}"
+preprocess_model = mlflow.pyfunc.load_model(model_uri=preprocess_model_uri)
+
+# COMMAND ----------
+
+X_train_processed = preprocess_model.predict(X_train)
+
+# COMMAND ----------
+
+X_train_processed_UC = mlflow.data.load_delta(table_name="lr_nhl_demo.dev.X_train_processed", version="0")
+X_train_processed_UC_PD = X_train_processed_UC.df.toPandas()
+
+X_train_processed_UC_PD
+
+# COMMAND ----------
+
+import mlflow
+from mlflow.models import Model, infer_signature, ModelSignature
+from mlflow.pyfunc import PyFuncModel
+from mlflow import pyfunc
+from hyperopt import hp, tpe, fmin, STATUS_OK, SparkTrials
+import pyspark.pandas as ps
+
+# Write X_train_processed to Unity Catalog for lineage tracking
+ps.from_pandas(X_train_processed).to_table("lr_nhl_demo.dev.X_train_processed", mode="overwrite")
+print("Writing X_train_processed to UC!")
+
+# Load a Unity Catalog table, train a model, and log the input table
+print("Loading X_train_processed_UC")
+X_train_processed_UC = mlflow.data.load_delta(table_name="lr_nhl_demo.dev.X_train_processed", version="0")
+X_train_processed_UC_PD = X_train_processed_UC.df.toPandas()
+
+import mlflow
+from sklearn.pipeline import Pipeline
+from lightgbm import LGBMRegressor
+from sklearn.linear_model import LinearRegression
+from sklearn.ensemble import RandomForestRegressor
+from xgboost import XGBRegressor
+from hyperopt import fmin, tpe, hp, STATUS_OK
+from hyperopt.pyll.base import scope
+from mlflow.models import Model
+from mlflow.pyfunc import PyFuncModel
+
+# Define the objective function to accept different model types
 def objective(params):
-    with mlflow.start_run(experiment_id="2824690123542843") as mlflow_run:
-        lgbmr_regressor = LGBMRegressor(**params)
+    model_type = params.pop('model_type')
 
-        model = Pipeline(
+    with mlflow.start_run(experiment_id="634720160613016") as mlflow_run:
+        mlflow.set_tag("model_type", model_type)
+        
+        # Select the model based on the model_type parameter
+        if model_type == 'lightgbm':
+            model = LGBMRegressor(**params)
+            print(f'Training LightGBM model with params: {params}')
+        elif model_type == 'linear':
+            model = LinearRegression()
+            print(f'Training LinearRegression model with params: {params}')
+        elif model_type == 'random_forest':
+            model = RandomForestRegressor(**params)
+            print(f'Training RandomForestRegressor model with params: {params}')
+        elif model_type == 'xgboost':
+            model = XGBRegressor(**params)
+            print(f'Training XGBRegressor model with params: {params}')
+        else:
+            raise ValueError(f'Unknown model type: {model_type}')
+        
+        # Create a pipeline with the selected model
+        pipeline = Pipeline(
             [
-                ("regressor", lgbmr_regressor),
+                ('regressor', model),
             ]
         )
 
@@ -590,30 +838,29 @@ def objective(params):
             silent=True,
         )
 
-        model.fit(
+        pipeline.fit(
             X_train_processed,
             y_train,
-            regressor__callbacks=[
-                lightgbm.early_stopping(5),
-                lightgbm.log_evaluation(0),
-            ],
-            regressor__eval_set=[(X_val_processed, y_val)],
+            # These callbacks are specific to LightGBM, so they should only be used for LightGBM
+            **({'regressor__callbacks': [lightgbm.early_stopping(5), lightgbm.log_evaluation(0)],
+                'regressor__eval_set': [(X_val_processed, y_val)]} if model_type == 'lightgbm' else {})
         )
-
+        
         # Log metrics for the training set
         mlflow_model = Model()
         pyfunc.add_to_model(mlflow_model, loader_module="mlflow.sklearn")
-        pyfunc_model = PyFuncModel(model_meta=mlflow_model, model_impl=model)
+        pyfunc_model = PyFuncModel(model_meta=mlflow_model, model_impl=pipeline)
         training_eval_result = mlflow.evaluate(
             model=pyfunc_model,
             data=X_train_processed.assign(**{str(target_col): y_train}),
             targets=target_col,
             model_type="regressor",
             evaluator_config={
-                "log_model_explainability": False,
-                "metric_prefix": "training_",
+                'log_model_explainability': False,
+                'metric_prefix': 'training_',
             },
         )
+        
         # Log metrics for the validation set
         val_eval_result = mlflow.evaluate(
             model=pyfunc_model,
@@ -621,11 +868,12 @@ def objective(params):
             targets=target_col,
             model_type="regressor",
             evaluator_config={
-                "log_model_explainability": False,
-                "metric_prefix": "val_",
+                'log_model_explainability': False,
+                'metric_prefix': 'val_',
             },
         )
-        lgbmr_val_metrics = val_eval_result.metrics
+        val_metrics = val_eval_result.metrics
+        
         # Log metrics for the test set
         test_eval_result = mlflow.evaluate(
             model=pyfunc_model,
@@ -633,42 +881,32 @@ def objective(params):
             targets=target_col,
             model_type="regressor",
             evaluator_config={
-                "log_model_explainability": False,
-                "metric_prefix": "test_",
+                'log_model_explainability': False,
+                'metric_prefix': 'test_',
             },
         )
-        lgbmr_test_metrics = test_eval_result.metrics
+        test_metrics = test_eval_result.metrics
 
-        loss = -lgbmr_val_metrics["val_r2_score"]
+        loss = -val_metrics["val_r2_score"]
 
         # Truncate metric key names so they can be displayed together
-        lgbmr_val_metrics = {
-            k.replace("val_", ""): v for k, v in lgbmr_val_metrics.items()
+        val_metrics = {
+            k.replace("val_", ""): v for k, v in val_metrics.items()
         }
-        lgbmr_test_metrics = {
-            k.replace("test_", ""): v for k, v in lgbmr_test_metrics.items()
+        test_metrics = {
+            k.replace("test_", ""): v for k, v in test_metrics.items()
         }
-
-        # # Add these lines to print the selected features
-        # feature_names = model.named_steps['preprocessor'].get_feature_names_out()
-        # selected_features = model.named_steps['feature_selection'].get_feature_names_out()
-        # print("Selected features after pipeline:", selected_features)
-        # mlflow.log_param("selected_features", selected_features.tolist())
 
         return {
             "loss": loss,
             "status": STATUS_OK,
-            "val_metrics": lgbmr_val_metrics,
-            "test_metrics": lgbmr_test_metrics,
-            "model": model,
+            "val_metrics": val_metrics,
+            "test_metrics": test_metrics,
+            "model": pipeline,
             "run": mlflow_run,
         }
 
 X_train_processed
-
-# COMMAND ----------
-
-
 
 # COMMAND ----------
 
@@ -693,19 +931,45 @@ X_train_processed
 
 # COMMAND ----------
 
-space = {
-    "colsample_bytree": 0.6587254729574785,
-    "lambda_l1": 0.18965542130830132,
-    "lambda_l2": 1.9402892441576718,
-    "learning_rate": 0.02517876184843777,
-    "max_bin": 29,
-    "max_depth": 8,
-    "min_child_samples": 140,
-    "n_estimators": 287,
-    "num_leaves": 166,
-    "subsample": 0.7501639626721491,
-    "random_state": 729986891,
-}
+from hyperopt.pyll.base import scope
+
+# Define the search space including the model type
+space = hp.choice('classifier_type', [
+    {
+        'model_type': 'lightgbm',
+        "colsample_bytree": hp.uniform("colsample_bytree", 0.5, 1.0),
+        "lambda_l1": hp.loguniform("lambda_l1", -5, 0),
+        "lambda_l2": hp.loguniform("lambda_l2", -5, 2),
+        "learning_rate": hp.loguniform("learning_rate", -5, -1),
+        "max_bin": scope.int(hp.quniform("max_bin", 20, 100, 1)),
+        "max_depth": scope.int(hp.quniform("max_depth", 3, 15, 1)),
+        "min_child_samples": scope.int(hp.quniform("min_child_samples", 20, 200, 1)),
+        "n_estimators": scope.int(hp.quniform("n_estimators", 100, 500, 1)),
+        "num_leaves": scope.int(hp.quniform("num_leaves", 31, 255, 1)),
+        "subsample": hp.uniform("subsample", 0.5, 1.0),
+        "random_state": 729986891,
+    },
+    # {
+    #     'model_type': 'linear',
+    # },
+    {
+        'model_type': 'random_forest',
+        "n_estimators": scope.int(hp.quniform("rf_n_estimators", 100, 500, 1)),
+        "max_depth": scope.int(hp.quniform("rf_max_depth", 3, 15, 1)),
+        "min_samples_split": hp.uniform("rf_min_samples_split", 0.1, 1.0),
+        "min_samples_leaf": hp.uniform("rf_min_samples_leaf", 0.1, 0.5),
+        "random_state": 729986891,
+    },
+    {
+        'model_type': 'xgboost',
+        "n_estimators": scope.int(hp.quniform("xgb_n_estimators", 100, 500, 1)),
+        "max_depth": scope.int(hp.quniform("xgb_max_depth", 3, 15, 1)),
+        "learning_rate": hp.loguniform("xgb_learning_rate", -5, -1),
+        "subsample": hp.uniform("xgb_subsample", 0.5, 1.0),
+        "colsample_bytree": hp.uniform("xgb_colsample_bytree", 0.5, 1.0),
+        "random_state": 729986891,
+    }
+])
 
 # COMMAND ----------
 
@@ -725,17 +989,14 @@ space = {
 
 # COMMAND ----------
 
-from hyperopt import SparkTrials
 trials = SparkTrials()
 
-# COMMAND ----------
-
-# trials = Trials()
+# Run the optimization
 fmin(
-    objective,
+    fn=objective,
     space=space,
     algo=tpe.suggest,
-    max_evals=1,  # Increase this when widening the hyperparameter search space.
+    max_evals=100,
     trials=trials,
 )
 
@@ -855,12 +1116,14 @@ if shap_enabled:
     ).fillna(mode)
 
     # Sample some rows from the validation set to explain. Increase the sample size for more thorough results.
-    example = X_val_processed.sample(n=min(100, X_val_processed.shape[0]), random_state=729986891).fillna(
-        mode
-    )
+    example = X_val_processed.sample(
+        n=min(100, X_val_processed.shape[0]), random_state=729986891
+    ).fillna(mode)
 
     # Use Kernel SHAP to explain feature importance on the sampled rows from the validation set.
-    predict = lambda x: model.predict(pd.DataFrame(x, columns=X_train_processed.columns))
+    predict = lambda x: model.predict(
+        pd.DataFrame(x, columns=X_train_processed.columns)
+    )
     explainer = KernelExplainer(predict, train_sample, link="identity")
     shap_values = explainer.shap_values(example, l1_reg=False, nsamples=500)
     summary_plot(shap_values, example)
@@ -902,7 +1165,14 @@ if shap_enabled:
 # COMMAND ----------
 
 # model_uri for the generated model
-print(f"runs:/{ mlflow_run.info.run_id }/model")
+best_model_uri = f"runs:/{ mlflow_run.info.run_id }/model"
+print(best_model_uri)
 
 # COMMAND ----------
+
+dbutils.jobs.taskValues.set(key="best_model_uri", value=best_model_uri)
+print(f'Successfully set the best_model_uri to {best_model_uri}')
+
+# COMMAND ----------
+
